% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.Approximate.R
\name{Analysis.Approximate}
\alias{Analysis.Approximate}
\title{Approimate Analysis}
\usage{
Analysis.Approximate(l = NULL, image.dir = "/Users/TKelly/Dropbox/Images/",
  hash = digest(runif(1), algo = "crc32"))
}
\arguments{
\item{l}{A list of the data sets to analyze. For each plot, the data saved in the files specified by l will be loaded from <./data/Model-...> and plotted.}

\item{image.dir}{The prefered output directory for the generated images. Include the final '/' on linux/mac and '\' on windows.}

\item{hash}{The code used for the unique identification of the composite images. Default is CRC32 hash of l.}
}
\description{
This script is designed analyze the set of runs <l> against the approimate equations of the model.
}
\examples{
Analysis.Approximate('Test01')

}

