% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.Convergence.R
\name{Analysis.Convergence}
\alias{Analysis.Convergence}
\title{Convergence Analysis}
\usage{
Analysis.Convergence(l = NULL, image.dir = "/Users/TKelly/Dropbox/Images/",
  hash = digest(runif(1), algo = "crc32"))
}
\arguments{
\item{l}{A list of the data sets to analyze. For each plot, the data saved in the files specified by l will be loaded and plotted.}

\item{image.dir}{The prefered output directory for the generated images. Include the final '/' on linux/mac and '\' on windows.}

\item{hash}{The code used for the unique identification of the composite images. Default is CRC32 hash of l.}
}
\description{
This script is designed analyze the set of runs <l> for convergence (i.e. stationary solution).
}
\examples{
Analysis.Convergence()
}

