% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.Correlation.R
\name{Analysis.Correlation}
\alias{Analysis.Correlation}
\title{Correlation Analysis}
\usage{
Analysis.Correlation(l = NULL, image.dir = "/Users/TKelly/Dropbox/Images/",
  hash = digest(runif(1), algo = "crc32"))
}
\arguments{
\item{l}{A list of the data sets to analyze. For each plot, the data saved in the files specified by l will be loaded and plotted.}

\item{hash}{The code used for the unique identification of the composite images. Default is CRC32 hash of l.}

\item{dir}{The prefered output directory for the generated images. Include the final '/' on linux/mac and '\' on windows.}
}
\description{
This script is designed to run and generate the analyses based on correlation between flows in the solution sets.
}
\examples{
Analysis.Correlation()
}

