% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Analysis.Raw.R
\name{Analysis.Raw}
\alias{Analysis.Raw}
\title{Raw Analysis}
\usage{
Analysis.Raw(l = NULL, image.dir = "/Users/TKelly/Dropbox/Images/",
  hash = digest(runif(1), algo = "crc32"))
}
\arguments{
\item{l}{A list of the data sets to analyze. For each plot, the data saved in the files specified by l will be loaded and plotted.}

\item{dir}{The prefered output directory for the generated images. Include the final '/' on linux/mac and '\' on windows.}
}
\description{
This script is designed to run and generate the analyses based on saved data.
}
\examples{
MCMC.Sample
}

