\name{CCELIM-package}
\alias{CCELIM-package}
\alias{CCELIM}
\docType{package}
\title{
\packageTitle{CCELIM}
}
\description{
\packageDescription{CCELIM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CCELIM}
\packageIndices{CCELIM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CCELIM}

Maintainer: \packageMaintainer{CCELIM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
