% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Analysis_Utils.R
\name{int_f}
\alias{int_f}
\title{int_f}
\usage{
int_f(x, mu1, mu2, sd1, sd2)
}
\arguments{
\item{x}{}

\item{mu1}{}

\item{mu2}{}

\item{sd1}{}

\item{sd2}{}
}
\description{
int_f
}
\examples{
in_f()
}
\keyword{Carlo,}
\keyword{Chain}
\keyword{MCMC,}
\keyword{Markov}
\keyword{Monte}
\keyword{mirror,}
\keyword{xsample,}

