% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CCELIM.R
\name{LIM}
\alias{LIM}
\title{LIM}
\usage{
LIM(name = "temp", iter = 1e+06, out.length = 10000,
  burn.length = 1e+05, jmp = 1.3, cycle = 1)
}
\arguments{
\item{name}{A list of the data sets to analyze. For each plot, the data saved in the files specified by l will be loaded and plotted.}

\item{iter}{The prefered output directory for the generated images. Include the final '/' on linux/mac and '\' on windows.}
}
\description{
This script is designed to run and generate the analyses based on saved data.
}
\examples{
LIM
}

