% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xsample.R
\name{MCMC.Sample}
\alias{MCMC.Sample}
\title{MCMC.sample()}
\usage{
MCMC.Sample(A = NULL, B = NULL, E = NULL, F = NULL, G = NULL,
  H = NULL, sdB = NULL, iter = 3000, outputlength = 100,
  burn.length = 1000, jmp = 1, x0 = NULL, test = TRUE, verbose = TRUE)
}
\arguments{
\item{A}{A matrix containing all the approximate equations.}

\item{B}{A vector containing all the measured values.}

\item{E}{A matrix containing all the exact equations.}

\item{F}{A vector containing the solutions to the exact equations.}

\item{G}{A matrix containing all the inequality relations.}

\item{H}{A vector containing the solutions to the inequality relations.}

\item{sdB}{The vector containing the uncertainty of the B vector in terms of SD. A minimum, relative uncertainty of 0.1 percent is applied to all values (i.e. B value of 10+/-0 will be changed to +/- 0.01).}

\item{iter}{The number of steps for the simulation to run for.}

\item{outputlength}{The total number of output solutions to save for analysis. A solution is saved very iter/outputlength steps.}

\item{burn.length}{The number of steps to take before starting to sample the solutions.}

\item{jmp}{A measurement of length taken between steps. Smaller values yield a higher percentage of excepted solutions but slower convergence. Aim for a value which yields an acceptance ratio of ~0.234}

\item{x0}{An inital solution to start from. Useful for restarting a simulation from a particular point.}

\item{test}{}

\item{verbose}{A boolean switch to turn on or off the progress counter.}
}
\description{
MCMC.sample()
}
\examples{
MCMC.Sample()
}
\keyword{Carlo,}
\keyword{Chain}
\keyword{MCMC,}
\keyword{Markov}
\keyword{Monte}
\keyword{mirror,}
\keyword{xsample,}

