% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIM_Utils.R
\name{norm}
\alias{norm}
\title{Norm}
\usage{
norm(x)
}
\arguments{
\item{A}{B E F G H}
}
\description{
This function allows you to express your love of cats.
}
\examples{
norm()
}
\keyword{Carlo,}
\keyword{Chain}
\keyword{MCMC,}
\keyword{Markov}
\keyword{Monte}
\keyword{mirror,}
\keyword{xsample,}

