% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notify.r
\name{notify}
\alias{notify}
\title{Notify}
\usage{
notify(name)
}
\arguments{
\item{name}{The name of the model run that is now completed.}
}
\description{
This function allows the automated notification of the completetion of the model.
}
\examples{
notify()
}

