% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFlows.R
\name{ReadFlows}
\alias{ReadFlows}
\title{ReadFlows}
\usage{
ReadFlows(model, solution,
  Flow.Worksheet = "/Users/TKelly/Dropbox/Documents/End2End2LayerCompartments.xlsx",
  flow.count = 30)
}
\arguments{
\item{Flow.Worksheet}{Location of the flow worksheet.}

\item{flow.count}{The number of flows to look for in the worksheet.}
}
\description{
This script is designed to run and generate the analyses based on saved data.
}
\examples{
ReadFlows()

}

