% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModel.R
\name{RunModel}
\alias{RunModel}
\title{RunModel()}
\usage{
RunModel(model, iter = 1e+05, out.length = 100, burn.length = 10000,
  jmp = 1)
}
\arguments{
\item{model}{The loaded dataframe structure generated by ReadModel().}

\item{iter}{The total number of steps for the simulation to take. Well have to try out several orders of magnitude to ensure convergence of the Markov Chain.}

\item{out.length}{The number of solutions to output for analysis. A solution is saved every iter/out.length steps.}

\item{jmp}{The relative distance between one step and the next. Larger the value the more likely a step to be rejected, and the smaller the step the more of them that will be required to reach convergence. (I adjust it to achive an acceptance ratio of ~0.234).}

\item{iter.burn}{The number of steps to take before sampling solutions for the final output. (Typically some fraction of iter.)}
}
\value{
solution of the xsample run as a dataframe.
}
\description{
Use this function to run the model that was generated by ReadModel().
}
\examples{
RunModel(model.A, iter=1e5, out.length=150, iter.burn=10000, jmp=2)
}
\keyword{MCMC}

