% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SaveData.R
\name{SaveModel}
\alias{SaveModel}
\title{SaveModel}
\usage{
SaveModel(model, name)
}
\arguments{
\item{model}{The dataframe generated by ReadModel()}

\item{name}{A unique name from the run.}
}
\description{
SaveModel
}
\examples{
cat_function()
}
\keyword{MCMC}

