% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveData.R
\name{SaveSolution}
\alias{SaveSolution}
\title{SaveSolution}
\usage{
SaveSolution(solution, name)
}
\arguments{
\item{solution}{The dataframe created by RunModel()}

\item{name}{A unique name from the run.}
}
\description{
SaveSolution
}
\examples{
SaveSolution()

}
\keyword{CCELIM}

