% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SaveData.R
\name{SaveSpread}
\alias{SaveSpread}
\title{SaveSpread}
\usage{
SaveSpread(model, solution, name)
}
\arguments{
\item{model}{The dataframe generated by ReadModel().}

\item{solution}{The dataframe created by RunModel()}

\item{name}{A unique name from the run.}
}
\description{
SaveSpread
}
\examples{
cat_function()
}
\keyword{CCELIM}

