% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIM_Utils.R
\name{SSR}
\alias{SSR}
\title{SSR}
\usage{
SSR(q1, q2, a, b)
}
\arguments{
\item{q1}{The prior solution}

\item{q2}{The proposed solution}

\item{a}{The matrix}

\item{b}{The goal solution}
}
\description{
This function calculates the difference of the sum of the squared residuals.
}
\examples{
SSR()

}
\keyword{Carlo,}
\keyword{Chain}
\keyword{MCMC,}
\keyword{Markov}
\keyword{Monte}
\keyword{mirror,}
\keyword{xsample,}

