% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_Utils.R
\name{int_f}
\alias{int_f}
\title{int_f}
\usage{
int_f(x, mu1, mu2, sd1, sd2)
}
\description{
Used to determine the difference of two normally distributed sets based on their count, mean and SD.
}
\examples{
in_f()

}
\keyword{Carlo,}
\keyword{Chain}
\keyword{MCMC,}
\keyword{Markov}
\keyword{Monte}
\keyword{mirror,}
\keyword{xsample,}

