% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCELIM.R
\name{LIM}
\alias{LIM}
\title{LIM}
\usage{
LIM(name = "temp", iter = 1e+06, out.length = 10000,
  burn.length = 1e+05, jmp = 1.3, cycle = 1)
}
\arguments{
\item{name}{A list of the data sets to analyze. For each plot, the data saved in the files specified by l will be loaded and plotted.}

\item{iter}{The prefered output directory for the generated images. Include the final '/' on linux/mac and '\' on windows.}

\item{out.length}{The number of desired solutions to be saved.}

\item{burn.length}{The number of steps to take before running the model for real (i.e. saving solutions).}

\item{jmp}{The length of step to take between iterations of the model (should aim for an acceptance ratio of ~0.234)}

\item{cycle}{The Data set to be used for the running of the model.}
}
\description{
A top-level entrance point for running the model. Primarily designed as an example due to the lack of flexability when calling this script. Can be used to automate the running of the model.
}
\examples{
LIM()

}

