% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mirror.R
\name{mirror}
\alias{mirror}
\title{Mirror
This function is taken, nearly unaltered, from the limSolve package which can be found on CRAN.}
\usage{
mirror(q1, g, h, k = length(q), jmp)
}
\description{
Mirror
This function is taken, nearly unaltered, from the limSolve package which can be found on CRAN.
}
\examples{
mirror()

}
\keyword{Carlo,}
\keyword{Chain}
\keyword{MCMC,}
\keyword{Markov}
\keyword{Monte}
\keyword{mirror,}
\keyword{xsample,}

