% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIM_Utils.R
\name{norm}
\alias{norm}
\title{Norm}
\usage{
norm(x)
}
\arguments{
\item{A}{B E F G H}
}
\description{
This function calculates X * X.
}
\examples{
norm()

}
\keyword{Carlo,}
\keyword{Chain}
\keyword{MCMC,}
\keyword{Markov}
\keyword{Monte}
\keyword{mirror,}
\keyword{xsample,}

