% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Notify.r
\name{notify}
\alias{notify}
\title{Grazing Analysis}
\usage{
notify(name)
}
\arguments{
\item{l}{A list of the data sets to analyze. For each plot, the data saved in the files specified by l will be loaded and plotted.}

\item{dir}{The prefered output directory for the generated images. Include the final '/' on linux/mac and '\' on windows.}
}
\description{
This script is designed analyze the set of runs <l> against the approimate equations of the model.
}
\examples{
Analysis.Approximate()
}

