% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LIM_Utils.R
\name{SSR}
\alias{SSR}
\title{SSR}
\usage{
SSR(q1, q2, a, b)
}
\arguments{
\item{q}{a b}
}
\description{
This function allows you to express your love of cats.
}
\examples{
SSR()
}
\keyword{Carlo,}
\keyword{Chain}
\keyword{MCMC,}
\keyword{Markov}
\keyword{Monte}
\keyword{mirror,}
\keyword{xsample,}

